move tauri utils to a separate modules (otherwise tests fail to run)
