+++ src/lib/util/util.ts
@@ -1,14 +1,7 @@
-import {
-  Dialog,
-  DialogChainObject,
-  LocalStorage,
-} from "quasar"
+import { Dialog, DialogChainObject } from "quasar"
 import { Component } from "vue"
 import * as process from "process"
-import { invoke } from "@tauri-apps/api/tauri"
 import { ApiPromise, WsProvider } from "@polkadot/api"
-import { appData } from "./appData"
-import { config } from "./appConfig"
 import { generateSlug } from "random-word-slugs"
 
 export interface IUtil {
@@ -18,7 +11,6 @@ export interface IUtil {
 }
 
 const nodeNameMaxLength = 64
-export const appName: string = process.env.APP_NAME || "subspace-desktop"
 
 export const random = (min: number, max: number): number =>
   Math.floor(Math.random() * (max - min)) + min
@@ -37,28 +29,6 @@ export async function showModal(
   })
 }
 
-export async function getLogPath(): Promise<string> {
-  return await invoke("custom_log_dir", { id: appName })
-}
-
-export async function errorLogger(error: unknown): Promise<void> {
-  if (error instanceof Error) {
-    const message = error.message
-    await invoke("frontend_error_logger", { message })
-  } else if (typeof error === "string") {
-    await invoke("frontend_error_logger", { message: error })
-  }
-}
-
-export async function infoLogger(info: unknown): Promise<void> {
-  if (info instanceof Error) {
-    const message = info.message
-    await invoke("frontend_info_logger", { message })
-  } else if (typeof info === "string") {
-    await invoke("frontend_info_logger", { message: info })
-  }
-}
-
 // TODO: rethink about the signature of this function, it should be string instead, and refactor the codebase accordingly
 export function toFixed(num: number, precision: number): number {
   if (!num) return 0
@@ -68,13 +38,6 @@ export function toFixed(num: number, precision: number): number {
 export function plotTimeMsEstimate(gb: number): number {
   return gb * 5e4
 }
-
-export async function resetAndClear(): Promise<void> {
-  await LocalStorage.clear()
-  await appData.clearDataDir()
-  await config.remove()
-}
-
 
 export function formatMS(duration: number): string {
   duration /= 1000
@@ -142,4 +105,14 @@ export function generateNodeName(): string {
     nodeName = slug + "-" + num.toString()
   } while (nodeName.length > nodeNameMaxLength)
   return nodeName
+}
+
+export function getErrorMessage(error: unknown): string | undefined {
+  if (error instanceof Error) {
+    return error.message;
+  } else if (typeof error === "string") {
+    return error;
+  } else {
+    return;
+  }
 }
